fix(packages, vault): sync FE ABIs with vault-contracts-aave-v4 (#1454)
* fix(packages, vault): sync FE ABIs with vault-contracts-aave-v4

* fix(vault): add missing params

+++ packages/babylon-ts-sdk/src/tbv/core/contracts/abis/BTCVaultRegistry.abi.ts
@@ -305,18 +305,18 @@ export const BTCVaultRegistryABI = [
       },
       {
         name: "universalChallengersVersion",
-        type: "uint32",
-        internalType: "uint32",
+        type: "uint16",
+        internalType: "uint16",
       },
       {
         name: "appVaultKeepersVersion",
-        type: "uint32",
-        internalType: "uint32",
+        type: "uint16",
+        internalType: "uint16",
       },
       {
         name: "offchainParamsVersion",
-        type: "uint32",
-        internalType: "uint32",
+        type: "uint16",
+        internalType: "uint16",
       },
       {
         name: "verifiedAt",
